Consistent exclusion for contig validation
During make_examples, there is a validation step to be sure contigs
reasonably overlap. This excludes some contigs, like chrM and extra
chromosomes but does it inconsistently. The contigs get excluded from
the list to use but then not during validation. This leads to errors on
small test datasets (bcbio has a chr22/chrM dataset that exposes this,
chrM is removed and then only 50% of the bases overlap so it fails).