Cleanup group_variants function in the haplotype_labeler so that it now groups the candidate and truth variants together.
This change is needed for two reasons.

First, by grouping our candidates and truth variants together we ensure that we determine the status each candidate and truth variant exactly once. Since we collect all of the truth variants in the region once, and then group them together, we can actually identify which truth variants were completely missed. This is a necessary feature to actually compute valid statistics about the quality of the labeling itself.

Two, we were potentially double counting truth variants previous, since we'd group our candidates up and then query for the truth variants nearby, allowing different groups to "validate" against the same truth candidate multiple times. This is likely extremely rare, but worth dealing with.

This CL produces better results than HEAD, in that we label more variants with the same FP rate and no runtime impact.

PiperOrigin-RevId: 191812678