Update make_examples to optionally write out labeling metrics to a text-format protobuf.
Add tests to make sure this works when provided and when not provided. Update variant_labeler to provide a labeler.metrics property in the baseclass that is None by default.

Only use candidate variants that overlap the confident_regions when we compute the labeling statistics.

Updates create_golden to write out a golden labeling output in the single shard case, and testdata to provide a constant for the golden labeling output.

PiperOrigin-RevId: 192307410