window_selector now counts overlapping candidate events as independent evidence for a window. Previously if we had an insertion of length N and a mismatch within N bases of the insertion, we'd only generate a single piece of evidence for assembling at the mismatch, despite having a +1 to the count from the insertion and a +1 from the mismatch. Now we allow these pieces of evidence to stack up, so we'd end up with a +2 here.
The value of this change is that it simplifies the logic of the window selector, as not counting overlapping events requires us to process counts per read. Whereas if everything is independent we can treat each position as independent, making it far easier to use the AlleleCounter to do our window selection.

PiperOrigin-RevId: 193380091