Update training code to warm start from a checkpoint. Previously in the estimator code, the pretrained checkpoint wasn't connected correctly either. After this CL, the default will be to train from model pretrained checkpoint.
Note that this implementation always ignore the last layer (related to number of classes). Previous implementation passes in num_classes and only chop that one off if the num_classes are different. We can change that later if that turns out to be important. (We'll check that when we train a WES model.)

PiperOrigin-RevId: 199528628