Reduce the default indel fraction for triggering candidate generation from 12% to 6%. Currently, the ref support is done per base. So, for Indel alleles, we tend to overcount the ref support.
After changing `vsc_min_fraction_indels` from 0.12 to 0.06 (without retraining), here is the F1 change:
* WES: Indel 96.6611% --> 96.7013%; SNP 0.999110% --> 99.9110%
* WGS: Indel 99.7427% --> 99.7950%; SNP 0.999622% --> 99.9623%
* PACBIO(chr20): Indel 98.3751% --> 98.3271%; SNP 99.8683% --> 99.8622%

Here is the TRUTH.FN changes for Indel:
* WES: 117 --> 115
* WGS: 1,485 --> 1,102
* PACBIO(chr20): 180 --> 172

Here is the QUERY.FP changes for Indel:
* WES: 55 --> 55
* WGS: 943 --> 837
* PACBIO(chr20): 156 --> 175

Here is the TRUTH.TP changes for Indel:
* WES: 2,483 --> 2,485
* WGS: 463,279 --> 463,662
* PACBIO(chr20): 9,992 --> 10,000

Here is the runtime change:
* WES:
 - make_examples: 8m18.258s --> 8m16.316s
 - call_variants: 1m50.501s --> 2m7.873s
* WGS:
 - make_examples: 76m40.676s --> 72m30.766s
 - call_variants: 173m3.542s --> 181m45.581s
* PACBIO:
 - make_examples: 65m42.630s --> 80m32.275s
 - call_variants: 182m14.369s --> 334m52.409s

Here is the changes of #entries in VCF (including PASS and RefCall):
* WES: 55,783 --> 57,204 (2.55% increase)
* WGS: 5,609,211 --> 5,807,515 (3.54% increase)
* PACBIO: 5,812,788 --> 10,446,991 (79.72% increase)

Based on these information, we'll use 6% for Illumina data (WGS and WGS). However, it seems like a good idea to still keep 12% for PacBio data. I will change the default for that in a next code change.

PiperOrigin-RevId: 273332119