Filter regions in make_examples to those containing candidates from a VCF if one is provided.
In both training and when using VcfCandidateImporter, only candidates contained in the provided VCF (--truth_variants or --proposed_variants) are retained. This change adds a precomputed filter to eliminate the regions without any candidates before spending the time and compute to preprocess them. This makes a small speedup in make_examples for training in general cases, but can be a huge difference when the VCF is sparse, such as when using VcfCandidateImporter for training or inference on structural variants (currently experimental).

PiperOrigin-RevId: 308341348