Move some pip install logic to run-prereq.sh instead of in Dockerfile.
This makes it possible to build the source code on a machine with:
```
DV_OPENVINO_BUILD=1 ./build-prereq.sh && ./build_and_test.sh
```
and have the right dependencies installed on the machine.

But in order to use `--use_openvino` with built binaries, the users still need to run:
```
  export LD_LIBRARY_PATH=/opt/intel/openvino/inference_engine/lib/intel64/:/opt/intel/openvino/inference_engine/external/tbb/lib/:/opt/intel/openvino/deployment_tools/ngraph/lib/${LD_LIBRARY_PATH:+:$LD_LIBRARY_PATH} \
      PYTHONPATH=/opt/intel/openvino/python/python3.6/:/opt/intel/openvino/deployment_tools/model_optimizer${PYTHONPATH:+:$PYTHONPATH}
```
in their environment.

PiperOrigin-RevId: 345248323