Add the remaining logic for `hp_tag_for_assembly_polishing`.
When `hp_tag_for_assembly_polishing` is used, it should be set to either 1 or 2, indicating which HP we want to call for assembly polishing. We want to put the HP-to-call on the top of the image, and we want to use the same coloring for the HP-to-call. (We used this symmetry when we train the model)

PiperOrigin-RevId: 350221439