Add a unit test for CallsFromVcf to understand the behavior when querying proposed VCF with contig names not in the VCF (entries or header).
Currently, if the querying range has a contig that is not in the header or the content of the proposed VCF, CallsFromVcf will crash.

But, this behavior might be too strict. This use case might happen if the user writes a VCF with proposed candidates but didn't cover all contigs in the BAM/FASTA. For example, if the user wants to only write a proposed VCF with just chr1, and only listed chr1 on the header -- which is reasonable because the user is not proposing any candidates outside chr1. But, right now this usage will crash in the CallsFromVcf function when querying this proposed VCF file.

I'm adding this unit test to confirm this issue, and will plan to relax this in the next code change.

PiperOrigin-RevId: 351194338