Unify DV and DT versions of RegionProcessor and make_examples_runner.
Between DV and DT, there are two RegionProcessor classes and two make_examples_runner functions, but they have diverged significantly, so here I make changes to both to arrive at a best-of-both-worlds implementation that works with a set of N samples.
As soon as this change is submitted, I'll remove one of the now-identical copies so DeepTrio can rely on the DeepVariant version.

PiperOrigin-RevId: 378916142