Exit earlier if users add --use_openvino when running run_deepvariant.py.
This is because we no longer enable OpenVINO by default in our Docker images. If users add --use_openvino and run with run_deepvariayt.py, they will only notice the issue until make_examples finishes, which is not ideal. (For example, https://github.com/google/deepvariant/issues/597 )

I also removed the relevant code for DV_OPENVINO_BUILD. Because we don't build with OpenVINO by default, we won't be testing the installation this time.

PiperOrigin-RevId: 508183064