Update inference_deepvariant.sh to copy over model.example_info.json, and update run_deepvariant.py to correctly check for model.example_info.json for customized models.
Add a `--customized_model_json` flag in inference_deepvariant.sh, so that when we use run_case_study.sh and when we use `--customized_model`, we don't need to make sure the gs:// model directory has a model.example_info.json file. We can simply supply a --customized_model_json flag. (This is flag is not designed for production use. For production models, it'll be better to have the `model.example_info.json` in the directory)

PiperOrigin-RevId: 780198342